+++ src/frontend/src/features/notifications/components/Toast.tsx
@@ -25,7 +25,13 @@ export function Toast({ state, ...props }: Readonly<ToastProps>) {
     <StyledToastContainer {...toastProps} ref={ref}>
       <StyledToast>
         <div {...contentProps}>{props.toast.content?.message}</div>
-        <Button square size="sm" invisible {...closeButtonProps}>
+        <Button
+          square
+          size="sm"
+          variant="ghost"
+          invisible
+          {...closeButtonProps}
+        >
           <RiCloseLine color="currentColor" />
         </Button>
       </StyledToast>