+++ src/frontend/src/features/recording/components/RequestRecording.tsx
@@ -47,10 +47,10 @@ export const RequestRecording = ({
   return (
     <div
       className={css({
-        backgroundColor: 'neutral.50',
+        backgroundColor: 'muted',
         borderRadius: 'surface',
         border: '1px solid',
-        borderColor: 'neutral.200',
+        borderColor: 'border',
         paddingY: '1rem',
         paddingX: '1rem',
         display: 'flex',